+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Tabbing from entry search directly to first entry
   - Tabbing from tag search to tag name input
   - Tabbing through entries skips the edit and show buttons
+  - Use arrow keys to reach those buttons
 
 
 ## 0.1.0