+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
+## unreleased
+
+- Improve keyboard navigation
+  - Tabbing from entry search directly to first entry
+  - Tabbing from tag search to tag name input
+  - Tabbing through entries skips the edit and show buttons
+  - Use arrow keys to reach those buttons
+
+
 ## 0.1.0
 _2026-07-21_